# bookstore-api/src/main/java/com/poc/bookstore/dto/CartItemRequestDto.java
package com.poc.bookstore.dto;

public class CartItemRequestDto {

	private Long bookId;
	private int quantity;

	public Long getBookId() {
		return bookId;
	}

	public void setBookId(Long bookId) {
		this.bookId = bookId;
	}

	public int getQuantity() {
		return quantity;
	}

	public void setQuantity(int quantity) {
		this.quantity = quantity;
	}

}